% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{chain}
\alias{chain}
\title{chain pairs}
\usage{
chain(pairs, chain_length = 4)
}
\arguments{
\item{pairs}{data.frame containing pairs in columns 1 and 2}

\item{chain_length}{Integer.}
}
\description{
chain pairs
}
