% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_pairs.R
\name{combine_pairs}
\alias{combine_pairs}
\title{Combine two sets of pairs into one data frame, randomising the chains}
\usage{
combine_pairs(pairs1, pairs2)
}
\arguments{
\item{pairs1}{data frame containing the variable \code{chain_number}}

\item{pairs2}{data frame containing the variable \code{chain_number}}
}
\value{
Data frame with rows from \code{pairs1} and \code{pairs2} randomised,
  with chains intact.
}
\description{
If \code{pairs1} and \code{pairs2} have any chain numbers in common,
\code{pairs2$chain_number} will be increased, so that chain numbers are
unique.
}
