% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_examples.R
\docType{data}
\name{data_standard}
\alias{data_standard}
\title{Example data set}
\format{A data frame with 3 variables. Column headers are 'media', 'core'
  & 'score'.
\describe{
  \item{media}{the label of the performance or media
  item.  Often a file name.}
  \item{core}{1 = core, 0 = non-core.  Core is for making a new scale,
  non-core is for locating on an existing scale.}
  \item{score}{score of the performance if available. May be missing.}
  }}
\usage{
data(data_standard)
}
\description{
Use with \code{pairs_generate}. All media are 'core'.
Used in projects where a new scale is to be created.
}
\keyword{datasets}
