% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicates}
\alias{duplicates}
\title{Show all duplicates in a set of pairs}
\usage{
duplicates(pairs, message = TRUE)
}
\arguments{
\item{pairs}{A data frame containing pairs, for example, an object returned
from \code{pairs_generate}.  This data frame must include the
\code{combinations} variable.}

\item{message}{Logical.  Should a message be returned if there are no
repeated pairs.}
}
\value{
A data frame showing which pairs are repeated, if any, or \code{NULL}
  if there are no repeated pairs.  A message is also optionally returned if
  there are no repeated pairs.  The column \code{pairs_i} contains the row
  numbers from \code{pairs} which have duplicated pairs.
}
\description{
Use this function on objects returned from \code{pairs_generate}, or similar
functions, to check if there are repeated comparisons for judges.
}
\examples{
# check if there are any duplicates in a set of pairs
gp <- pairs_generate(data_standard,
                     av_inclusions = 10,
                     inclusion_tolerance = 2,
                     chain_length = 1)
duplicates(gp)

# check if there are any duplicates for the judges
allocated_pairs <- pairs_allocate(gp, judges = 5)
judge_sets <- split(allocated_pairs, f = as.factor(allocated_pairs$judge))
judge_duplicates <- lapply(judge_sets, duplicates)

}
