% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{find_duplicate_pairs}
\alias{find_duplicate_pairs}
\title{Find duplicate comparisons in judgement df, given that 'left' and
'right' may be reversed}
\usage{
find_duplicate_pairs(x)
}
\arguments{
\item{x}{data frame of pairs, containing \code{left} and \code{right}}
}
\description{
Find duplicate comparisons in judgement df, given that 'left' and
'right' may be reversed
}
