% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs2.R
\name{inclusions}
\alias{inclusions}
\title{Tabulate inclusions for each performance}
\usage{
inclusions(pairs)
}
\arguments{
\item{pairs}{A data frame containing pairs of performances.}

\item{return}{A named vector of performance inclusions}
}
\description{
Tabulate inclusions for each performance
}
