% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersperse_chains.R
\name{mix_chains}
\alias{mix_chains}
\title{Randomise Chains}
\usage{
mix_chains(x)
}
\arguments{
\item{x}{Data frame of pairs, including the variable \code{chain_number}.}
}
\description{
Randomise the row order in a set of pairs, but keep the chaining. Useful for
combining standard pairs with core v non-core pairs.
}
