% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_examples.R
\docType{data}
\name{NAPLAN_example}
\alias{NAPLAN_example}
\title{Example student performance lables}
\format{Character vector.}
\usage{
data(NAPLAN_example)
}
\description{
A mock example resembling labels of student performances used in the equating
of NAPLAN Writing, a large scale assessment for Australian primary and
secondary school students.
}
\keyword{datasets}
