% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs_all.R
\name{pairs_all}
\alias{pairs_all}
\title{Generate exhaustive pairs.}
\usage{
pairs_all(scripts, reps = 1)
}
\arguments{
\item{scripts}{Character vector of performances to compare.}

\item{reps}{Integer value at least 1. Use this to make import file for
multiple judges.  Use reps value at least equal to the number of judges.}
}
\value{
A data frame containing columns "left", "right" and "chain_number".
  "chain_number" can be read as judge number for the purposes of import into
  Pairwise Comparisons Application.
}
\description{
\code{all_pairs} generates all unique pairs for pairwise comparisons.
}
\details{
Exhaustive pairs (each media item paired with all others) can be created for
using any number of judges.  Each judge will view all pairs. Allocate
n*(n-1)/2 pairs to each judge.
}
\examples{
  all_pairs(letters)

}
