% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate.R
\name{pairs_allocate}
\alias{pairs_allocate}
\title{Allocate pairs to judges}
\usage{
pairs_allocate(pairs, judges)
}
\arguments{
\item{pairs}{A data frame containing pairs in two columns, such as returned
from \code{pairs_generate}}

\item{judges}{Single integer specifying the number of judges to allocate
pairs evenly to, or a vector of integers specifying the number of pairs to
allocate to each judge.}
}
\value{
The input data frame with a new column \code{judge} added.
}
\description{
Allocate pairs to judges
}
\examples{
gp <- pairs_generate(letters,
                   av_inclusions = 10,
                   inclusion_tolerance = 1,
                   chain_length = 1)
allocated_pairs <- pairs_allocate(gp, judges = 5)

}
