% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{pairs_export}
\alias{pairs_export}
\title{Export pairs to .csv}
\usage{
pairs_export(gp, file)
}
\arguments{
\item{gp}{A data frame returned from \code{pairs_generate}, or other pair
generating function.}

\item{file}{Character string naming a file for writing to.}
}
\description{
\code{pairs_export} selects the variables \code{left}, \code{right} and
\code{chain_number} from \code{gp}, and writes to file using
\code{write.csv}.
}
\seealso{
\code{\link{pairs_generate}}
}
