% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs_generate_cnc.R
\name{pairs_generate_cnc}
\alias{pairs_generate_cnc}
\title{Pair non-core performances with core performances}
\usage{
pairs_generate_cnc(
  media,
  nc_include,
  chain_length = 4,
  separation_constraint = FALSE
)
}
\arguments{
\item{media}{A data frame containing the variables: \code{media}, a character
vector of performance labels; \code{core}, a numeric vector, whether core
(\code{1}) or non-core (\code{0}); and \code{score}, a numeric vector of
scores.}

\item{nc_include}{Integer.  The number of core performances to pair with each
non-core performance.}

\item{chain_length}{Integer.  The number of successive pairs to contain a
common performance (used to increase the efficiency of judging).}

\item{separation_constraint}{Numeric, or \code{FALSE} for no separation
constraint.  The maximum absolute difference between performance scores in
a pair.}
}
\value{
A data frame containing the variables \code{left}, \code{right}, and
  \code{chain_number}.  Each \code{left, right} pair will contain one "core"
  and one "non-core" performance.
}
\description{
This type of pair generation is for placing new performances on an existing
scale.  "Core" performances should have location estimates, and "non-core"
performances are to be placed on the scale formed using the "core"
performances.
}
\examples{
pairs_generate_cnc(data_cnc[1:4, ], nc_include = 1, chain_length = 1)

}
\references{
Humphry, S., & Bredemeyer, K. (2022). Pairwise comparison scale
  extension using core linking sets. Frontiers in Education, 7, [826742].
  \url{https://doi.org/10.3389/feduc.2022.826742}
}
