% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{pairs_inclusion_range}
\alias{pairs_inclusion_range}
\title{Range of inclusions per media/performance}
\usage{
pairs_inclusion_range(pairs)
}
\arguments{
\item{pairs}{\code{data.frame} returned from \code{pairs_generate},
\code{pairs_generate_cnc}, or similar function.}
}
\value{
Numeric vector length 2, reporting the minimum and maximum number of
  inclusions per performance in the set of pairs used as its first argument.
}
\description{
Range of inclusions per media/performance
}
\examples{
pairs <- pairs_generate(letters[1:3],
                        av_inclusions = 2,
                       inclusion_tolerance = 1,
                       chain_length = 1)
pairs_inclusion_range(pairs)
}
