% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{pairs_plot}
\alias{pairs_plot}
\title{Plot the frequency of unique pairs included in \code{pairs}}
\usage{
pairs_plot(pairs)
}
\arguments{
\item{pairs}{A data frame returned from \code{pairs_generate} or similar
function.}
}
\description{
Plot the frequency of unique pairs included in \code{pairs}
}
\examples{
gp <- pairs_generate(letters[1:7],
                     av_inclusions = 15,
                     inclusion_tolerance = 2,
                     chain_length = 1)
pairs_plot(gp)
}
\seealso{
\code{\link{pairs_plot_inclusions}}
}
