% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{pairs_plot_inclusions}
\alias{pairs_plot_inclusions}
\title{barchart to see how well each script is represented.}
\usage{
pairs_plot_inclusions(pairs)
}
\arguments{
\item{pairs}{A data frame containing pairs of performances.}
}
\description{
barchart to see how well each script is represented.
}
