% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{pairs_plot_scores}
\alias{pairs_plot_scores}
\title{Plot the scores of performances in each pair}
\usage{
pairs_plot_scores(pairs, scores_df)
}
\arguments{
\item{pairs}{Data frame returned from \code{pairs_generate} or similar
function.}

\item{scores_df}{Data frame containing the variables \code{media, core,
score}}
}
\description{
This plot can be used to check that groups of performances are not isolated,
and provides an indication that the separation constraint used in
\code{pairs_generate} or similar function is working correctly.  Currently,
overplotting of performance labels occurs for performances which have close
scores.
}
\examples{
pairs <- pairs_generate(data_standard[1:10, ],
                      av_inclusions = 10,
                      inclusion_tolerance = 2, chain_length = 1)
pairs_plot_scores(pairs, data_standard[1:10, ])
}
