% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs_generate_is.R
\name{pairs_sample}
\alias{pairs_sample}
\title{Sample pairs from the set of all possible pairs}
\usage{
pairs_sample(media, av_inclusions, separation_constraint = NULL)
}
\arguments{
\item{media}{Character vector representing the names of the performances, or
a data frame containing the variables \code{media, core, score}.}

\item{av_inclusions}{Integer specifying the average number of times each
performance should be included in the generated pairs.}

\item{separation_constraint}{Numeric.  Maximum absolute difference in scores
between performances in a pair.}
}
\description{
Similar to \code{pairs_generate} but allows finer control.  All performances
are included in the set of generated pairs, but unlike \code{pairs_generate}
no attempt is made to conform to a range of inclusions per performance.  Use
with \code{improve_inclusions} to improve the range of inclusions of
performances in the set of pairs.
}
\details{
Combine \code{pairs_sample} in a workflow with \code{improve_inclusions},
\code{chain} and \code{switch_lr} to acheive the same result as
\code{pairs_generate}

\code{improve_inclusions} optionally shows an animation of inclusions and
optionally allows repeated pairs.
}
\examples{
pairs_sample(letters[1:5], av_inclusions = 2)
}
\seealso{
\code{\link{pairs_generate}}, \code{\link{improve_inclusions}},
  \code{\link{chain}} and  \code{\link{switch_lr}}.
}
