% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{plot_pairs}
\alias{plot_pairs}
\title{Frequency plot of pairs}
\usage{
plot_pairs(pairs)
}
\arguments{
\item{pairs}{data.frame returned from \code{pairs_make}.}
}
\description{
Frequency plot of pairs
}
