% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{reverse_duplicates}
\alias{reverse_duplicates}
\title{Find repeated pairs in reverse left/right order}
\usage{
reverse_duplicates(gp)
}
\arguments{
\item{gp}{data.frame returned from \code{pairs_make}.}
}
\value{
A data frame containing repeated pairs and their match.
}
\description{
Find repeated pairs in reverse left/right order
}
