% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pairs.R
\name{score_diff}
\alias{score_diff}
\title{Show scores and score difference for a set of pairs}
\usage{
score_diff(pairs, media)
}
\arguments{
\item{pairs}{Data frame returned from \code{generate_pairs} or similar
function}

\item{media}{Data frame containing the variables \code{media, core, score}.
\code{score} must be numeric.}
}
\value{
Data frame similar to \code{pairs}, but with \code{left_score},
 \code{right_score}, and \code{score_difference} columns added.
}
\description{
Use this function to check that a separation constraint has been successfully
applied using \code{pairs_generate}, or similar function.
}
\examples{
# check separation constraint
pairs <- pairs_generate(data_standard,
                        av_inclusions = 5,
                        inclusion_tolerance = 2,
                        chain_length = 1,
                        separation_constraint = 30)
pairs_scores <- score_diff(pairs, data_standard)
max(pairs_scores$score_difference)

}
