% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicates}
\alias{duplicates}
\title{Find Repeated Pairs}
\usage{
duplicates(pairs)
}
\arguments{
\item{pairs}{data.frame returned from \code{pairs_make}.}
}
\value{
A data frame containing repeated pairs and their match.
}
\description{
Find Repeated Pairs
}
