% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs_all.R
\name{pairs_all}
\alias{pairs_all}
\title{Generate exhaustive pairs}
\usage{
pairs_all(media, n_judges = 1, separation_constraint = NULL)
}
\arguments{
\item{media}{Character vector of performances to compare, or data frame
containing the variables \code{media, core, score}.}

\item{n_judges}{Integer, at least 1. Use this to make import file for
multiple judges.  Use a value at least equal to the number of judges.}

\item{separation_constraint}{Numeric. Maximum absolute difference in scores
for performances in a pair.  If a separation constraint is specified, which
is less than the maximum absolute difference between any two performances,
then the resulting design will not be exhaustive.}
}
\value{
A data frame containing columns \code{left}, \code{right} and \code{judge}. If
  you're importing into Pairwise Comparisons Application, change \code{judge} to
  \code{chain_number}.
}
\description{
\code{pairs_all} generates all unique pairs for pairwise comparisons.
}
\details{
Exhaustive pairs (each media item, or performance, paired with all others)
can be created for  any number of judges.  Each judge will view all
pairs. Allocate n.(n-1)/2 pairs to each judge, where n is the number of
performances.
}
\examples{
  pairs_all(letters[1:4])

}
