% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R, R/pairs_importcsv.R
\name{pairs_export}
\alias{pairs_export}
\alias{pairs_export}
\title{Export pairs to .csv}
\usage{
pairs_export(gp, file)

pairs_export(gp, file)
}
\arguments{
\item{gp}{data.frame returned from \code{pairs_make}}

\item{file}{Character string naming a file for writing}

\item{gp}{data.frame returned from \code{pairs_make}}

\item{file}{Character string naming a file for writing}
}
\description{
\code{pairs_export} selects the variables 'left', 'right', 'chain_number' and
writes to file using \code{write.csv}.

\code{pairs_export} selects the variables 'left', 'right', 'chain_number' and
writes to file using \code{write.csv}.
}
\seealso{
\code{\link{pairs_make}}
}
