% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R, R/pairs_importcsv.R
\name{pairs_import.csv}
\alias{pairs_import.csv}
\alias{pairs_import.csv}
\title{read in data}
\usage{
pairs_import.csv(scores_file, separation_constraint = NULL)

pairs_import.csv(scores_file, separation_constraint = NULL)
}
\arguments{
\item{scores_file}{a path and .csv file containing the label of the
performance ("media"), whether the performance is core (1) or non-core (0),
and the score of the performance, if available.}

\item{separation_constraint}{Logical.}

\item{scores_file}{a path and .csv file containing the label of the
performance ("media"), whether the performance is core (1) or non-core (2),
and the score of the performance, if available.}

\item{choose_core}{logical.  Is the core column to be used?}
}
\value{
A data.frame corresponding to the data read in. An error will be
  returned if the headers are not correct in \code{scores_file}, if there are
  missing media labels, if 'core' is not all 0 or 1.  An error will also be
  returned if \code{separation_constraint == TRUE} and there are missing or
  non-numeric scores.
}
\description{
\code{pairs_import.csv} reads in perforance data, and makes the data
available for other pairwise analysis functions.

\code{pairs_import.csv} reads in perforance data, and makes the data available for
other pairwise analysis functions.
}
\details{
\code{pairs_import.csv} uses \code{read.csv} to read a .csv file, and issues
warnings if it detects that the format of the input file is not suitable for
subsequent analysis.

Warnings will be issued if: 1) the headers in \code{scores_file} are not
"media", "core", "score", or 2) if \code{choose_coreNcore = TRUE} and the
"core" column does not contain only 1s and 0s, or 3) if
\code{separtion_constraint} is specified but the "score" column is not
numeric or not complete.

\code{pairs_import.csv} uses \code{read.csv} to read a .csv file, and issues warnings
if it detects that the format of the input file is not suitable for
subsequent analysis.

Warnings will be issued if: 1) the headers in \code{scores_file} are not
"media", "core", "score", or 2) if \code{choose_coreNcore = TRUE} and the
"core" column does not contain only 1s and 0s, or 3) if
\code{separtion_constraint} is specified but the "score" column is not
numeric or not complete.
}
