% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{pairs_plot_inclusions}
\alias{pairs_plot_inclusions}
\title{Barchart to see how well each performance is represented}
\usage{
pairs_plot_inclusions(pairs)
}
\arguments{
\item{pairs}{A data frame containing pairs of performances.}
}
\description{
To test if \code{pairs_generate} produces pairs which conform to the
specified range of inclusions of performances, use
\code{pairs_plot_inclusions} together with \code{abline} (see the example
below), or use \code{\link{pairs_inclusion_range}}.
}
\examples{
pairs <- pairs_generate(letters[1:10],
                         av_inclusions = 10,
                         inclusion_tolerance = 2,
                         chain_length = 1)
pairs_plot_inclusions(pairs)
abline(h = c(10-2, 10+2))
}
