% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pairs_make.R
\name{switch_lr}
\alias{switch_lr}
\title{randomize left/right presentation}
\usage{
switch_lr(x)
}
\arguments{
\item{x}{data frame containing paired performances in cols 1 and 2.}
}
\description{
This function switches the value of the column 1 with the value of column 2,
for random row numbers.
}
